4. Defining the Delete Function This function is responsible for removing a snippet from the database.
A. Receiving the Identifier
The function takes an identifier as input, which specifies which snippet should be deleted.
B. Executing the Delete Query
Prisma is used to delete the snippet from the database based on the provided identifier.
C. Revalidating the Cache
The homepage cache is refreshed to remove the deleted snippet from the list of available snippets.
D. Redirecting the User
After deletion, the user is redirected to the homepage.