2. Explanation of Caching and Dynamic Behavior By default, Next.js caches data fetched inside a server component. However, this behavior can be changed using dynamic settings.
A. Disabling Caching Using Dynamic Mode (Commented Out)
A configuration is provided to force the route to be dynamic, ensuring fresh data is always fetched.
This is done by disabling automatic caching, making every request fetch fresh data from the database.
B. Alternative Method: Disabling Revalidation
Another approach is explicitly setting the revalidation time to zero.
This means that when the site is built, the data will not be cached and will always be dynamically fetched.