Imports PostList: A component responsible for rendering a list of posts. It accepts a prop fetchData, which is a function that fetches posts based on a specific query. This component is imported from @/components/posts/post-list.
fetchPostBySearch: A function from @/lib/query/post, which is responsible for querying the database or an API to fetch posts based on a search term (term). This function is used to get posts relevant to the search term provided in the query.

React: Required for defining the functional component SearchPage.