Dialog and Form Layout Dialog: The entire form is wrapped in a Dialog component, which is used to create a modal pop-up.
The modal is triggered by clicking the "New topic" button, which is wrapped in a DialogTrigger component.

DialogContent: This component defines the modal's content, where the form resides. The modal is set to have a maximum width of 425px for better responsiveness.

DialogHeader: Contains the title and description for the modal.

DialogTitle: The title of the modal, "Create a topic."

DialogDescription: A short description about the form, instructing users to write a new topic and save when done.