Form Fields Name Input:
Label: Labels the name input field as "Name."

Input: The actual input field where the user can type the name of the topic. It is controlled by the name attribute (name="name") to match the form data.

Error Handling: If the createTopics action returns an error for the name field, the error message will be displayed below the input field in red.

Description Textarea:

Label: Labels the description input field as "Description."

Textarea: A text area where the user can write a description of the topic.

Error Handling: If there are errors related to the description, they will be displayed under the textarea in red text.