Imports PostList: A component that renders a list of posts. It takes a function (fetchData) as a prop that fetches posts based on certain criteria, such as by topic slug.
PostCreateForm: A form component for creating a new post. It accepts a slug prop, which is the identifier for the topic under which the new post will be created.

fetchPostbyTopicSlug: A function imported from @/lib/query/post. This function fetches posts related to a specific topic, identified by the topic’s slug.

React: The core React library for building user interfaces. It’s used to define the component structure and manage the component lifecycle.