Post List <PostList fetchData={() => fetchPostbyTopicSlug(slug)} />:
The PostList component is responsible for displaying the list of posts related to the topic.

The fetchData prop is passed a function that calls fetchPostbyTopicSlug(slug), which fetches posts filtered by the topic’s slug.

This means that when the PostList component is rendered, it will display all posts associated with the topic specified by the slug.