Dialog UI:
The dialog window is triggered when the user clicks the "New Post" button (inside the DialogTrigger component).

Dialog Content: Contains the form with two fields:

Title Field: An input field where the user provides the title of the post.

Content Field: A textarea where the user provides the content of the post.

Both fields have associated labels to provide context to the user.