Imports:
prisma: This import allows access to the Prisma ORM instance, which interacts with the database. The prisma object is used to query the database for post data.

notFound: This function is imported from next/navigation and is used to handle scenarios where the requested data (in this case, a post) cannot be found. It triggers a 404 error page when the post is not found.

React: Standard import of React to define the component.