Fetching Post Data:
The PostShow component is an asynchronous function, indicated by the async keyword.

Inside the component, it uses await new Promise((res) => setTimeout(res, 3000)); to introduce a 3-second delay before fetching the post. This is likely used for simulating loading or introducing a delay for testing purposes.

Fetching the Post:

It then fetches the post from the database using prisma.post.findFirst(). The findFirst method is used to retrieve the first post that matches the given criteria, in this case, a post with the ID equal to postId.

If no post is found with that ID, it calls notFound(), which triggers a 404 error response, rendering a "not found" page.