Rendering the Post:
If the post is found, the component renders:

The post title inside an <h1> element with a font-bold, my-2, and text-2xl Tailwind CSS classes for styling.

The post content inside a <p> element with a border, rounded, and p-4 class for styling (adding padding and a border around the content).